Reworked TestWatcher to be synchronous (#3)
So fsnotify has a chance to pass at least part of test-cases.

The test was reworked from asynchronous event-sourcing to
make-action-await-event synchronous fashion - the previous one caused
fsnotify to not discover half of the events, moreover it looked like
create/write/delete events on files within a directory can be preempted
by later one single delete event of that directory. Weird.

So now the focus would be on implementing FSEvents and seeing how it
would differ from fsnotify.