fix TestNotifyExample for OS X (#54)
Since FSEvents tends to send unrelated, historical flags altogether
with the actual event, its watcher implementation needs to bookkeep
them and filter out duplicated events before dispatching to the user.
The problem arises of the first event for a given path is Write -
FSEvents sends Create+Write but watcher has no knowledge what was the
real previous state of the event, that's why it will also dispatch
Create event to the user. This should be documented (#48) or even better
fixed.

That's why in test-cases we begin with dummy Create's to "initialize"
state of the history bookkeeping map. Lame, I know.