Switch to InotifyInit1 syscall for mips64/etc
The InotifyInit syscall doesn't exist on all architectures, but
InotifyInit1 does.  In the kernel implementation, inotify_init is just a
wrapper around inotify_init1 and passes in zero for the only parameter
which is a flags field.

In this change, the flag passed in is IN_CLOEXEC, which will close the
inotify fd if the process is forked and exec is called.  This was
requested by ppknap.