Handle pointers that are now integers in Go 1.10 (fixes #135)
The Core Services reference types became integer shaped after being
pointer shaped in Go 1.9. I moved the thing out into a constant zero or
nil variable according to the Go version, in order to avoid duplicating
a bunch of code into separate files.

This is the required fix according to
https://go-review.googlesource.com/c/go/+/66332.

This passes tests with Go 1.9.2. and 1.10beta1.